fix(common/query): portable-compiler fixes exposed by the full CI matrix
Pre-existing #290 issues my SDK-0.31 branch's full matrix surfaced:

- MSVC's traditional preprocessor mangles raw string literals containing
  quotes inside macro arguments (C2017/C3688); rewrite the two escaped-quote
  lexer vectors as ordinary escaped strings.
- Clang's -Wconversion implies -Wsign-conversion (GCC's does not), so the
  int-indexed token access all over pj_query failed the macOS build. Route
  vector indexing through per-class tok()/at()/slice() helpers that
  centralize the int → size_type cast; positions stay int end-to-end since
  they mirror Qt cursor offsets.

Also: drop toolbox_mosaico's now-unused <limits> include and
data_load_mp4's obsolete test include-directory (review nits).

+++ common/query/include/pj_query/edit.hpp
@@ -60,7 +60,8 @@ struct CursorContext {
 // Returns token index, or -1 if cursor is in whitespace.
 [[nodiscard]] inline int findActiveToken(const std::vector<Token>& tokens, int cursor) {
   for (int index = 0; index < static_cast<int>(tokens.size()); ++index) {
-    if (cursor >= tokens[index].start && cursor <= tokens[index].end) {
+    if (cursor >= tokens[static_cast<std::size_t>(index)].start &&
+        cursor <= tokens[static_cast<std::size_t>(index)].end) {
       return index;
     }
   }
@@ -79,7 +80,7 @@ struct CursorContext {
 
   ctx.token_index = findActiveToken(tokens, cursor);
   if (ctx.token_index >= 0) {
-    ctx.active_token = tokens[ctx.token_index];
+    ctx.active_token = tokens[static_cast<std::size_t>(ctx.token_index)];
   }
 
   // Slice the stream at the cursor: everything up to and including the active
